pyproject.toml
[project]
name = "sf-data-pipelines-quant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "numpy>=2.3.4",
    "polars>=1.35.1",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "wrds>=3.4.0",
]
